fix: moved the endpoint used for getting users by branch to UserController
